Refactor rooms to service
