--- backend/app/controllers/rooms.controller.js
@@ -1,88 +1,0 @@
-import fs from "fs";
-import uuid from "uuid/v4";
-import path from "path";
-import request from "request";
-
-const roomFilePath = "../file/matrix.room.web.json";
-
-const fetchFromFile = () => {
-  const roomFileExists = fs.existsSync(roomFilePath);
-  if (!roomFileExists) {
-    createRoomFileSync();
-  }
-
-  const roomsData = fs.readFileSync(roomFilePath);
-  const roomsDetail = JSON.parse(roomsData);
-
-  return new Promise(resolve => resolve(roomsDetail));
-};
-
-const createRoomFileSync = () => {
-  const roomsData = [];
-
-  roomsData[0] = {
-    id: uuid(),
-    name: "The Dock",
-    disableMeeting: true,
-  };
-
-  const niceNames = [
-    "Nebuchadnezzar",
-    "Logos",
-    "Osiris",
-    "Icarus",
-    "Caduceus",
-    "Brahma",
-    "Novalis",
-    "Vigilant",
-    "Zion",
-  ];
-
-  for (const niceName of niceNames) {
-    roomsData.push({
-      id: uuid(),
-      name: niceName
-    });
-  }
-
-  fs.mkdirSync(path.dirname(roomFilePath), { recursive: true });
-  fs.writeFileSync(roomFilePath, JSON.stringify(roomsData));
-};
-
-const fetchFromEnvironment = (env) => {
-  const roomsData = env.ROOMS_DATA;
-  const roomsDetail = JSON.parse(roomsData);
-
-  return new Promise(resolve => resolve(roomsDetail));
-};
-
-const fetchFromRemote = (env) => {
-
-  let url = env.ROOMS_DATA;
-
-  return new Promise(function(resolve, reject){
-        request(url, function (error, response, body) {
-            // in addition to parsing the value, deal with possible errors
-            if (error) return reject(error);
-            try {
-                // JSON.parse() can throw an exception if not valid JSON
-                resolve(JSON.parse(body));
-            } catch(e) {
-                reject(e);
-            }
-        });
-    });
-};
-
-const fetchRooms = (strategy) => {
-  switch (strategy) {
-    case "ENVIRONMENT":
-      return fetchFromEnvironment(process.env);
-    case "REMOTE":
-      return fetchFromRemote(process.env);
-    default:
-      return fetchFromFile();
-  }
-};
-
-export default fetchRooms;
